added views for mlparameters and featureshiddenlayers
